+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## [x]
 - Auto generate plugin readme script
 - Footer permanent link
-- Translate plugin SLA to english
 - Adjust new DevTools
 - Remove update plugin readme script
+- Adjust new GeneratePluginInfosHelper
 
 ## [2.3.3]
 - Fix